auren/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML dependencies
crewai==0.30.11
crewai-tools==0.2.6
openai>=1.13.3
sentence-transformers==2.3.1
chromadb==0.4.22
faiss-cpu==1.7.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0

# Computer Vision
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
# Note: dlib requires special installation, see README

# WhatsApp integration
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Utils
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3
